\name{CompareL1_L2_TV1}
\alias{CompareL1_L2_TV1}
\title{
   Compare L1, L2 and TV on a sparse signal.
}
\description{
   Compare L1, L2 and TV on a sparse signal.
}
\usage{
CompareL1_L2_TV1(N, M, per)
}
\arguments{
  \item{N}{
     Size of the sparse signal to generate , integer.
}
  \item{M}{
     Number of measurements.
}
  \item{per}{
      Percentage of spikes.
}
}
\author{
Mehmet Suzen
}
