\name{DFTMatrix0}
\alias{DFTMatrix0}
\title{
 Generate Discrete Fourier Transform Matrix using DFTMatrixPlain.
}
\description{
  Generate Discrete Fourier Transform Matrix (NxN).
}
\usage{
DFTMatrix0(N)
}
\arguments{
  \item{N}{
   Integer value determines the dimension of the square matrix.
}
}
\value{
It returns a NxN square matrix.
}
\author{
Mehmet Suzen
}

\seealso{
 DFTMatrixPlain
}
\examples{
 DFTMatrix0(2)
}
