\name{DFTMatrixPlain}
\alias{DFTMatrixPlain}
\title{
  Generate Plain Discrete Fourier Transform Matrix without the coefficient
}
\description{
  Generate plain Discrete Fourier Transform Matrix (NxN) without a coefficient.
}
\usage{
DFTMatrixPlain(N)
}
\arguments{
  \item{N}{
  Integer value defines the dimension of the square plain DFT matrix.
}
}
\value{
It returns a NxN square matrix.
}
\author{
Mehmet Suzen
}

\examples{
DFTMatrixPlain(2)
}
