\name{GaussianMatrix}
\alias{GaussianMatrix}
\title{
 Generate Gaussian Random Matrix
}
\description{
 Generate Gaussian Random Matrix ( zero mean and standard deviation one.)
}
\usage{
GaussianMatrix(N, M)
}
\arguments{
  \item{N}{
   Integer value determines number of rows.
}
  \item{M}{
   Integer value determines number of columns.
}
}
\value{
Returns  MxN matrix.
}
\author{
Mehmet Suzen
}

\examples{
GaussianMatrix(3,2)

}
