% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Lnorm.R
\name{Lnorm}
\alias{Lnorm}
\title{L-p norm of a given complex vector}
\usage{
Lnorm(X, p)
}
\arguments{
\item{X,}{a complex vector, can be real too.}

\item{p,}{norm value}
}
\value{
L-p norm of the complex vector
}
\description{
L-p norm of a given complex vector
}
\author{
Mehmet Suzen
}

