% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objectiveL1.R
\name{objectiveL1}
\alias{objectiveL1}
\title{Objective function for ridge L1 penalty}
\usage{
objectiveL1(x, T, phi, y, lambda)
}
\arguments{
\item{x,}{unknown vector}

\item{T,}{transform bases}

\item{phi,}{measurement vector}

\item{lambda,}{penalty term}
}
\description{
Objective function for ridge L1 penalty
}
\note{
Thank you Jason Xu of Washington University for pointing out complex number handling
}
\author{
Mehmet Suzen
}

