% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/objectiveL2.R
\name{objectiveL2}
\alias{objectiveL2}
\title{Objective function for Tikhinov L2 penalty}
\usage{
objectiveL2(x, T, phi, y, lambda)
}
\arguments{
\item{x,}{unknown vector}

\item{T,}{transform bases}

\item{phi,}{measurement vector}

\item{lambda,}{penalty term}
}
\description{
Objective function for Tikhinov L2 penalty
}
\note{
Thank you Jason Xu of Washington University for pointing out complex number handling
}
\author{
Mehmet Suzen
}

